requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
pandas==2.3.1
openpyxl==3.1.5
python-dateutil==2.9.0 